requirements.txt
flask
flask-script
flask-sqlalchemy
flask-migrate
flask-wtf
flask-bcrypt
flask-openid
flask-oauth
flask-login
flask-principal
flask-debugtoolbar
flask-mongoengine
flask-restful
flask-celery-helper
celery
flower
psycopg2
ipython